add: binance and kraken exchange integration
- Add Binance and Kraken API clients to Balance class
- Aggregate balances across on-chain, Binance, and Kraken sources
- Add support for new assets: LINK, BNB, PAXG
- Update portfolio targets config
- Add python-binance and krakenex to requirements
- Allow float percent values for set_target command
- Fix deploy.sh shebang and volume mount

+++ utils/command_handlers.py
@@ -48,7 +48,7 @@ async def get_targets(update: Update, context: ContextTypes.DEFAULT_TYPE) -> Non
 async def set_target(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
     try:
         symbol = context.args[0].upper()
-        percent = int(context.args[1])
+        percent = float(context.args[1])
         portfolio.set_target(symbol, percent)
         write_json(targets_file_path, portfolio.get_targets())
